modificações na autenticação
